+++ index.js
@@ -4,7 +4,7 @@
 
 var util = require('util');
 var Orchestrator = require('orchestrator');
-var chalk = require('chalk');
+var gutil = require('gulp-util');
 
 // format orchestrator errors
 var formatError = function(e) {
@@ -19,33 +19,27 @@ function Gulp(){
 
   // Logging
   this.on('task_start', function(e){
-    gulp.log('Running', "'"+chalk.cyan(e.task)+"'...");
+    gutil.log('Running', "'"+gutil.colors.cyan(e.task)+"'...");
   });
   this.on('task_stop', function(e){
-    gulp.log('Finished', "'"+chalk.cyan(e.task)+"' in "+chalk.magenta(e.duration)+" seconds");
+    var time = gutil.prettyTime(e.duration);
+    gutil.log('Finished', "'"+gutil.colors.cyan(e.task)+"'", "in", gutil.colors.magenta(time.value), time.shortUnit);
   });
 
   this.on('task_err', function(e){
     var msg = formatError(e);
-    gulp.log('Errored', "'"+chalk.cyan(e.task)+"' in "+chalk.magenta(e.duration)+" seconds "+chalk.red(msg));
+    var time = gutil.prettyTime(e.duration);
+    gutil.log('Errored', "'"+gutil.colors.cyan(e.task)+"'", "in", gutil.colors.magenta(time.value), time.shortUnit, gutil.colors.red(msg));
   });
 }
 util.inherits(Gulp, Orchestrator);
-
-Gulp.prototype.log = function(){
-  if (this.env.silent) return;
-  var sig = '['+chalk.green('gulp')+']';
-  var args = Array.prototype.slice.call(arguments);
-  args.unshift(sig);
-  console.log.apply(console, args);
-  return this;
-};
 
 Gulp.prototype.taskQueue = Gulp.prototype.seq;
 Gulp.prototype.task = Gulp.prototype.add;
 Gulp.prototype.run = function(){
   var tasks = Array.prototype.slice.call(arguments);
   
+  // impose our opinion of "default" tasks onto orchestrator
   if (!tasks.length) {
     tasks = ['default'];
   }
@@ -56,7 +50,6 @@ Gulp.prototype.src = require('./lib/createInputStream');
 Gulp.prototype.dest = require('./lib/createOutputStream');
 Gulp.prototype.watch = require('./lib/watchFile');
 
-Gulp.prototype.createGlobStream = require('glob-stream').create;
 Gulp.prototype.formatFile = require('./lib/formatFile');
 Gulp.prototype.bufferFile = require('./lib/bufferFile');
 Gulp.prototype.streamFile = require('./lib/streamFile');